Rename checkYourAnswersSections
We’ll use this in the read-only view, so this naming is more
appropriate. I’ve also moved it to sit in the applications
subfolder
