+++ server/utils/assessments/utils.ts
@@ -19,7 +19,7 @@ import paths from '../../paths/assess'
 import { TasklistPageInterface } from '../../form-pages/tasklistPage'
 import Assess from '../../form-pages/assess'
 import { UnknownPageError } from '../errors'
-import { embeddedSummaryListItem } from '../checkYourAnswersUtils'
+import { embeddedSummaryListItem } from '../applications/summaryListUtils'
 import reviewSections from '../reviewUtils'
 import Apply from '../../form-pages/apply'
 import { kebabCase, linkTo } from '../utils'
